+++ web/vtctld2/src/app/dashboard/keyspaceView/keyspaceView.component.html
