+++ web/vtctld2/src/app/dashboard/keyspaceView/keyspaceView.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { ROUTER_DIRECTIVES, ActivatedRoute, Router } from '@angular/router';
-import { KeyspaceService } from '../keyspaceService/keyspace.service';
+import { KeyspaceService } from '../../shared/keyspaceService/keyspace.service';
 import { MD_CARD_DIRECTIVES } from '@angular2-material/card';
 import { MD_BUTTON_DIRECTIVES } from '@angular2-material/button';
